[tests] Move typemap rewrite coverage to net11
Move the cross-component generated typemap rewrite tests out of the net10 Build.Tests project now that the typemap generator targets net11. Keep the same assertions in the existing net11 integration test project, which references both components directly.

Fixes the solution compilation regression introduced by #12631.
